requirements.txt
django==4.1.7
djangorestframework==3.14.0
mysqlclient==2.1.1
opencv-contrib-python==4.7.0.72
face_recognition==1.3.0
imutils==0.5.4